refactor: chat tools
