+++ templates/types/streaming/nextjs/app/components/ui/chat/chat-message-content.tsx
@@ -1,13 +1,10 @@
 import {
   ChatMessage,
   ContentPosition,
-  getAnnotationData,
-  Message,
-  MessageAnnotation,
   useChatMessage,
 } from "@llamaindex/chat-ui";
-import { JSONValue } from "ai";
-import ChatTools from "./chat-tools";
+import {ChatMarkdown} from "./chat-markdown";
+import { ToolAnnotations } from "./chat-tools";
 
 export function ChatMessageContent() {
   const { message } = useChatMessage();
@@ -17,27 +14,10 @@ export function ChatMessageContent() {
       component: <ToolAnnotations message={message} />,
     },
   ];
-  return <ChatMessage.Content content={customContent} />;
-}
-
-type ToolData = {
-  toolCall: {
-    id: string;
-    name: string;
-    input: {
-      [key: string]: JSONValue;
-    };
-  };
-  toolOutput: {
-    output: JSONValue;
-    isError: boolean;
-  };
-};
-
-function ToolAnnotations({ message }: { message: Message }) {
-  const annotations = message.annotations as MessageAnnotation[] | undefined;
-  const toolData = annotations
-    ? (getAnnotationData(annotations, "tools") as unknown as ToolData[])
-    : null;
-  return toolData?.[0] ? <ChatTools data={toolData[0]} /> : null;
+  return (
+    <ChatMessage.Content
+      content={customContent}
+      markdownComponent={ChatMarkdown}
+    />
+  );
 }
